# Conflicts:
#	app/src/main/java/ru/practicum/android/diploma/search/presentation/SearchVacanciesViewModel.kt

+++ app/src/main/java/ru/practicum/android/diploma/filter/domain/models/Settings.kt
@@ -4,7 +4,7 @@ data class Settings(
     val industry: Industry,
     val country: Country,
     val area: Area,
-    val salary: Int,
+    val salary: Long,
     val onlyWithSalary: Boolean,
     val filterOn: Boolean
 )
